+++ src/app/constants/Module.ts
@@ -1,0 +1,3 @@
+export enum Module {
+    CONFIG_CONTROL = 'configControl',
+}